Moved concerns to prevent autoloading conflict
