+++ app/controllers/token_authenticate_me/api/v1/invites_controller.rb
@@ -1,10 +1,10 @@
-require 'token_authenticate_me/controllers/invitable'
+require 'token_authenticate_me/concerns/controllers/invitable'
 
 module TokenAuthenticateMe
   module Api
     module V1
       class InvitesController < BaseController
-        include TokenAuthenticateMe::Controllers::Invitable
+        include TokenAuthenticateMe::Concerns::Controllers::Invitable
 
         model TokenAuthenticateMe::Invite
         serializer TokenAuthenticateMe::InviteSerializer
