+++ app/controllers/token_authenticate_me/api/v1/password_resets_controller.rb
@@ -1,10 +1,10 @@
-require 'token_authenticate_me/controllers/password_resetable'
+require 'token_authenticate_me/concerns/controllers/password_resetable'
 
 module TokenAuthenticateMe
   module Api
     module V1
       class PasswordResetsController < BaseController
-        include TokenAuthenticateMe::Controllers::PasswordResetable
+        include TokenAuthenticateMe::Concerns::Controllers::PasswordResetable
       end
     end
   end
