+++ app/controllers/token_authenticate_me/api/v1/sessions_controller.rb
@@ -1,10 +1,10 @@
-require 'token_authenticate_me/controllers/sessionable'
+require 'token_authenticate_me/concerns/controllers/sessionable'
 
 module TokenAuthenticateMe
   module Api
     module V1
       class SessionsController < BaseController
-        include TokenAuthenticateMe::Controllers::Sessionable
+        include TokenAuthenticateMe::Concerns::Controllers::Sessionable
       end
     end
   end
