+++ app/controllers/token_authenticate_me/api/v1/users_controller.rb
@@ -1,11 +1,11 @@
-require 'token_authenticate_me/controllers/token_authenticateable'
+require 'token_authenticate_me/concerns/controllers/token_authenticateable'
 
 module TokenAuthenticateMe
   module Api
     module V1
       class UsersController < BaseController
         include ApiMe
-        include TokenAuthenticateMe::Controllers::TokenAuthenticateable
+        include TokenAuthenticateMe::Concerns::Controllers::TokenAuthenticateable
 
         skip_before_action :authenticate, only: [:create]
