+++ app/models/token_authenticate_me/invite.rb
@@ -1,9 +1,9 @@
-require 'token_authenticate_me/models/invitable'
+require 'token_authenticate_me/concerns/models/invitable'
 
 
 module TokenAuthenticateMe
   class Invite < ActiveRecord::Base
-    include TokenAuthenticateMe::Models::Invitable
+    include TokenAuthenticateMe::Concerns::Models::Invitable
 
     def accept!(current_user)
       # no-op, override for app specific accept logic
