+++ app/models/token_authenticate_me/session.rb
@@ -1,8 +1,8 @@
 require 'token_authenticate_me'
-require 'token_authenticate_me/models/sessionable'
+require 'token_authenticate_me/concerns/models/sessionable'
 
 module TokenAuthenticateMe
   class Session < ActiveRecord::Base
-    include TokenAuthenticateMe::Models::Sessionable
+    include TokenAuthenticateMe::Concerns::Models::Sessionable
   end
 end
