+++ app/models/token_authenticate_me/user.rb
@@ -1,8 +1,8 @@
-require 'token_authenticate_me/models/authenticatable'
+require 'token_authenticate_me/concerns/models/authenticatable'
 
 module TokenAuthenticateMe
   class User < ActiveRecord::Base
-    include TokenAuthenticateMe::Models::Authenticatable
+    include TokenAuthenticateMe::Concerns::Models::Authenticatable
 
     def self.policy_class
       TokenAuthenticateMe::UserPolicy
