+++ lib/token_authenticate_me/concerns/models/invitable.rb
@@ -1,0 +1,20 @@
+require 'active_support/concern'
+
+module TokenAuthenticateMe
+  module Concerns
+    module Models
+      module Invitable
+        extend ActiveSupport::Concern
+
+        included do
+          belongs_to :creator, class_name: User
+
+          validates :requestor, presence: true
+          validates :email, presence: true
+
+          default_scope { order('created_at ASC') }
+        end
+      end
+    end
+  end
+end
